cleaned up files and updated readme
