+++ clt/models/activations.py
@@ -1,5 +1,5 @@
 import torch
-from typing import Optional
+from typing import Optional, Tuple
 
 
 class BatchTopK(torch.autograd.Function):
@@ -44,7 +44,6 @@ class BatchTopK(torch.autograd.Function):
         x_flat = x.reshape(-1)  # [B*F_total]
         ranking_flat = ranking_tensor_to_use.reshape(-1)
 
-        indices: Optional[torch.Tensor] = None
         if k_total_batch > 0:
             # Global top-k on the flattened ranking tensor
             _, flat_indices = torch.topk(ranking_flat, k_total_batch, sorted=False)
@@ -95,3 +94,117 @@ class BatchTopK(torch.autograd.Function):
             # This is just illustrative.
 
         return grad_input, None, None, None  # Gradients for k, straight_through, x_for_ranking are None
+
+
+class TokenTopK(torch.autograd.Function):
+    @staticmethod
+    def forward(
+        ctx, x: torch.Tensor, k: float, straight_through: bool, x_for_ranking: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        """Applies TokenTopK activation (TopK per token/row).
+
+        Args:
+            x: Input tensor of shape [B_tokens, F_total].
+            k: The number or fraction of top elements to keep per token.
+               If k < 1, it's treated as a fraction of F_total.
+               If k >= 1, it's treated as an integer count.
+            straight_through: Whether to use straight-through estimator for gradients.
+            x_for_ranking: Optional tensor to use for ranking. If None, x is used.
+
+        Returns:
+            Output tensor with TokenTopK applied.
+        """
+        B_tokens, F_total = x.shape
+
+        if F_total == 0:  # Handle empty feature dimension
+            if straight_through:
+                ctx.save_for_backward(torch.zeros_like(x, dtype=torch.bool))
+            return torch.zeros_like(x)
+
+        k_per_token: int
+        if 0 < k < 1:
+            k_per_token = int(torch.ceil(torch.tensor(k * F_total)).item())
+        elif k >= 1:
+            k_per_token = int(k)
+        else:  # k <= 0
+            if straight_through:
+                ctx.save_for_backward(torch.zeros_like(x, dtype=torch.bool))
+            return torch.zeros_like(x)
+
+        # Clamp k_per_token to be at most F_total
+        k_per_token = min(k_per_token, F_total)
+
+        ranking_tensor_to_use = x_for_ranking if x_for_ranking is not None else x
+
+        if k_per_token > 0:
+            # Apply topk for each row (token)
+            # topk returns values and indices. We only need indices to create the mask.
+            _, topk_indices_per_row = torch.topk(ranking_tensor_to_use, k_per_token, dim=-1, sorted=False)
+
+            # Create mask
+            mask = torch.zeros_like(x, dtype=torch.bool)
+            # Use scatter_ to set True at the topk_indices for each row
+            mask.scatter_(-1, topk_indices_per_row, True)
+        else:  # k_per_token is 0
+            mask = torch.zeros_like(x, dtype=torch.bool)
+
+        if straight_through:
+            ctx.save_for_backward(mask)
+        else:
+            ctx.save_for_backward(mask)  # For now, non-STE also behaves like STE in backward
+
+        ctx.straight_through = straight_through
+
+        return x * mask.to(x.dtype)
+
+    @staticmethod
+    def backward(ctx, *args: torch.Tensor) -> tuple[torch.Tensor | None, None, None, None]:
+        """Backward pass for TokenTopK (identical to BatchTopK's STE path)."""
+        if len(args) != 1:
+            raise ValueError(f"TokenTopK backward expected 1 gradient tensor, got {len(args)}")
+        grad_output = args[0]
+
+        if ctx.straight_through:
+            (mask,) = ctx.saved_tensors
+            grad_input = grad_output * mask.to(grad_output.dtype)
+        else:
+            (mask,) = ctx.saved_tensors
+            grad_input = grad_output * mask.to(grad_output.dtype)
+
+        return grad_input, None, None, None
+
+
+class JumpReLU(torch.autograd.Function):
+    @staticmethod
+    def forward(ctx, input: torch.Tensor, threshold: torch.Tensor, bandwidth: float) -> torch.Tensor:
+        ctx.save_for_backward(input, threshold)
+        ctx.bandwidth = bandwidth
+        return (input >= threshold).float() * input
+
+    @staticmethod
+    def backward(ctx, *grad_outputs: torch.Tensor) -> Tuple[Optional[torch.Tensor], Optional[torch.Tensor], None]:
+        grad_output = grad_outputs[0]
+        input, threshold = ctx.saved_tensors
+        bandwidth = ctx.bandwidth
+        needs_input_grad, needs_threshold_grad, _ = ctx.needs_input_grad
+
+        grad_input = None
+        grad_threshold = None
+
+        if needs_input_grad:
+            ste_mask = (input >= threshold).type_as(grad_output)
+            grad_input = grad_output * ste_mask
+
+        if needs_threshold_grad:
+            is_near_threshold = torch.abs(input - threshold) <= (bandwidth / 2.0)
+            local_grad_theta = (-input / bandwidth) * is_near_threshold.type_as(input)
+            grad_threshold_per_element = grad_output * local_grad_theta
+
+            if grad_threshold_per_element.dim() > threshold.dim():
+                dims_to_sum = tuple(range(grad_threshold_per_element.dim() - threshold.dim()))
+                grad_threshold = grad_threshold_per_element.sum(dim=dims_to_sum)
+                if threshold.shape != torch.Size([]):
+                    grad_threshold = grad_threshold.reshape(threshold.shape)
+            else:
+                grad_threshold = grad_threshold_per_element.sum()
+        return grad_input, grad_threshold, None
