+++ clt/training/utils.py
@@ -1,0 +1,13 @@
+import datetime
+
+
+# Helper function to format elapsed time
+def _format_elapsed_time(seconds: float) -> str:
+    """Formats elapsed seconds into HH:MM:SS or MM:SS."""
+    td = datetime.timedelta(seconds=int(seconds))
+    hours, remainder = divmod(td.seconds, 3600)
+    minutes, seconds = divmod(remainder, 60)
+    if td.days > 0 or hours > 0:
+        return f"{td.days * 24 + hours:02d}:{minutes:02d}:{seconds:02d}"
+    else:
+        return f"{minutes:02d}:{seconds:02d}"
